One collection for all users -> one user per collection
Got rid of a lot of chunky code resulted from embedded array of objects.
It's much fresher to be able to use the driver's native functions instead of using `updateOne()` for everything.

Feels like this change is well worth it, just like some other decisions previously. It has already made it much easier to do things (and more secure since some data isn't passed around), not to say it will sort and filter the data for me in the future.